Move authentication functions into separate middleware module
